% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiemRegress-package.R
\docType{package}
\name{RiemRegress-package}
\alias{RiemRegress-package}
\title{Regression Methods on Manifolds}
\description{
We provide a number of algorithms to estimate fundamental statistics including
}
