% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmreg.elocal.R
\name{rmreg.elocal}
\alias{rmreg.elocal}
\title{Extrinsic Local Regression on Manifold-Valued Data}
\usage{
rmreg.elocal(y, x = NULL, h = seq(from = 0.1, to = 2, by = 0.1),
  parallel = TRUE, predict = FALSE, x.predict = NULL)
}
\description{
Extrinsic Local Regression on Manifold-Valued Data
}
